Update code.js
I noticed an array set as the initial sum. I knew it was incorrect and changed it to 0 so that the initial sum would be set as an integer instead of an array. 
Plagiarism Statement:
“I certify that I have listed all sources used to complete this exercise, including the use
of any Large Language Models. All of the work is my own, except where stated
otherwise. I am aware that plagiarism carries severe penalties and that if plagiarism is
suspected, charges may be filed against me without prior notice.”